[FIX] #31 flyway 설정
- JPA와의 통합을 위한 DB 스키마 및 엔티티 일부 수정
- flyway 적용

+++ src/main/java/com/depromeet/yellowcardapi/domain/Drink.java
@@ -29,7 +29,7 @@ public class Drink {
     private String description;
 
     @Column
-    private Integer proof;
+    private Double proof;
 
     @Column
     private Integer price;
